+++ symtype.h
@@ -10,7 +10,7 @@ the GNU General Public License, version 2, 1991.
 ********************************************/
 
 /*
- * $MawkId: symtype.h,v 1.6 2009/12/13 19:36:24 Jonathan.Nieder Exp $
+ * $MawkId: symtype.h,v 1.7 2010/05/07 08:18:41 tom Exp $
  * @Log: symtype.h,v @
  * Revision 1.6  1996/02/01  04:39:43  mike
  * dynamic array scheme
@@ -178,7 +178,7 @@ void relocate_resolve_list(int, int, FBLOCK *, int, unsigned, int);
 
 /* hash.c */
 unsigned hash(const char *);
-unsigned hash2(const char *, unsigned);
+unsigned hash2(const char *, size_t);
 SYMTAB *insert(const char *);
 SYMTAB *find(const char *);
 const char *reverse_find(int, PTR);
